fix: improving leaders screens
