+++ src/pages/Leader/createLeader.tsx
@@ -1,13 +1,22 @@
 import React, { useState, useEffect } from 'react';
-import { KeyboardAvoidingView, ScrollView, StyleSheet, Platform } from 'react-native';
-import { useAuth } from '@hooks/Auth';
-import AppLoading from '@components/AppLoading';
-import FormLeader from '@components/Form/Leader';
-import Toast from '@components/Toast';
-import { separeDDDFromPhoneNumber } from '@utils/separeDDDfromPhoneNumber';
-import api from '@services/Api';
-import { AppColors, Leader } from '../../types';
+import { View, StyleSheet, KeyboardAvoidingView, ScrollView, Platform } from 'react-native';
+import { Card, TextInput, Text, useTheme } from 'react-native-paper';
 import { useNavigation } from '@react-navigation/native';
+import DateTimePicker from '@react-native-community/datetimepicker';
+import { TextInputMask } from 'react-native-masked-text';
+import Input from '@components/Input';
+import Form from '@components/Form';
+import Button from '@components/Button';
+import AppLoading from '@components/AppLoading';
+import Toast from '@components/Toast';
+import { useAuth } from '@hooks/Auth';
+import api from '@services/Api';
+import { Formik } from 'formik';
+import { format } from 'date-fns';
+import ptBR from 'date-fns/locale/pt-BR';
+import { separeDDDFromPhoneNumber } from '@utils/separeDDDfromPhoneNumber';
+import { Leader, AppColors } from '../../types';
+import { formSchema } from './utils/formValidation';
 
 interface SubmitFormData {
   tx_nome: string;
@@ -16,15 +25,20 @@ interface SubmitFormData {
   email: string;
   password: string;
   confirmPassword: string;
+  user_type: number;
 }
 
 const CreateLeader: React.FC = () => {
   const [loading, setLoading] = useState<boolean>(true);
+
+  const [showsDatePicker, setShowsDatePicker] = useState<boolean>(false);
+
   const [successToastVisible, setSuccessToastVisible] = useState<boolean>(false);
   const [errorToastVisible, setErrorToastVisible] = useState<boolean>(false);
 
   const auth = useAuth();
   const navigation = useNavigation();
+  const theme = useTheme();
 
   useEffect(() => {
     setLoading(false);
@@ -36,7 +50,7 @@ const CreateLeader: React.FC = () => {
       if (errorToastVisible) {
         setErrorToastVisible(false);
       }
-    }, 5000);
+    }, 2500);
 
     return () => {
       clearTimeout(timer);
@@ -50,7 +64,7 @@ const CreateLeader: React.FC = () => {
         setSuccessToastVisible(false);
         navigation.goBack();
       }
-    }, 5000);
+    }, 2500);
 
     return () => {
       clearTimeout(timer);
@@ -70,7 +84,7 @@ const CreateLeader: React.FC = () => {
       .post('/v1/leaders', {
         tx_nome,
         dt_nascimento,
-        email : email.toLowerCase(),
+        email: email.toLowerCase(),
         password,
         lider_id: auth.user.id,
         nu_ddd: dddPhoneNumber,
@@ -89,7 +103,148 @@ const CreateLeader: React.FC = () => {
   return (
     <ScrollView style={styles.container}>
       <KeyboardAvoidingView behavior={Platform.OS === 'ios' ? 'padding' : undefined} enabled>
-        <FormLeader leader={{} as Leader} onSubmit={onSubmit} />
+        <Formik
+          onSubmit={onSubmit}
+          validationSchema={formSchema}
+          initialValues={{
+            tx_nome: '',
+            dt_nascimento: new Date(),
+            nu_telefone: '',
+            email: '',
+            password: '',
+            confirmPassword: '',
+            user_type: 2, // 1 para admin | 2 para líder
+          }}
+        >
+          {({ values, touched, setFieldTouched, handleSubmit, handleChange, setFieldValue, handleBlur, errors }) => (
+            <React.Fragment>
+              <Form title="INFORMAÇÕES PESSOAIS">
+                <Card.Content style={styles.cardContent}>
+                  <Text style={styles.label}>Nome</Text>
+
+                  <Input
+                    value={values.tx_nome}
+                    placeholder="Nome completo"
+                    onChangeText={handleChange('tx_nome')}
+                    onBlur={() => setFieldTouched('tx_nome', true)}
+                    error={errors.tx_nome && touched.tx_nome ? true : false}
+                    theme={theme}
+                  />
+                </Card.Content>
+
+                <Card.Content style={styles.cardContent}>
+                  <Text style={styles.label}>Data de nascimento</Text>
+
+                  <Input
+                    disabled
+                    value={format(new Date(values.dt_nascimento), 'd - M - yyyy', { locale: ptBR })}
+                    error={errors.dt_nascimento ? true : false}
+                    style={{ width: '45%' }}
+                    right={<TextInput.Icon name="calendar-month-outline" onPress={() => setShowsDatePicker(true)} />}
+                    theme={theme}
+                  />
+
+                  {showsDatePicker && (
+                    <DateTimePicker
+                      value={new Date()}
+                      mode="date"
+                      is24Hour
+                      display="default"
+                      onTouchCancel={() => setShowsDatePicker(false)}
+                      onChange={(e, selectedDate) => {
+                        setShowsDatePicker(false);
+                        if (selectedDate) setFieldValue('dt_nascimento', selectedDate);
+                      }}
+                    />
+                  )}
+                </Card.Content>
+
+                <Card.Content>
+                  <Text style={styles.label}>Telefone</Text>
+
+                  <TextInput
+                    keyboardType="numeric"
+                    textContentType="telephoneNumber"
+                    placeholder="(__) _____  -  ____"
+                    error={errors.nu_telefone && touched.nu_telefone ? true : false}
+                    value={values.nu_telefone}
+                    onChangeText={handleChange('nu_telefone')}
+                    onBlur={() => setFieldTouched('nu_telefone', true)}
+                    style={[styles.input, { width: '65%', paddingLeft: 5 }]}
+                    mode="outlined"
+                    theme={theme}
+                    render={props => (
+                      // @ts-ignore
+                      <TextInputMask
+                        type={'cel-phone'}
+                        options={{
+                          maskType: 'BRL',
+                          withDDD: true,
+                          dddMask: '(99) ',
+                        }}
+                        {...props}
+                      />
+                    )}
+                  />
+                </Card.Content>
+              </Form>
+
+              <Form title="CRIAR CONTA">
+                <Card.Content style={styles.cardContent}>
+                  <Text style={styles.label}>E-mail</Text>
+
+                  <Input
+                    textContentType="emailAddress"
+                    keyboardType="email-address"
+                    placeholder="email@email.com"
+                    value={values.email}
+                    onChangeText={handleChange('email')}
+                    onBlur={() => setFieldTouched('email', true)}
+                    error={errors.email && touched.email ? true : false}
+                    theme={theme}
+                  />
+                </Card.Content>
+
+                <Card.Content style={styles.cardContent}>
+                  <Text style={styles.label}>Senha</Text>
+                  <Text style={[styles.label, { color: 'rgb(168, 168, 168)', fontStyle: 'italic' }]}>
+                    5 ou mais caracteres
+                  </Text>
+
+                  <Input
+                    value={values.password}
+                    onBlur={() => setFieldTouched('password', true)}
+                    onChangeText={handleChange('password')}
+                    secureTextEntry={true}
+                    placeholder="*********"
+                    error={errors.password && touched.password ? true : false}
+                    theme={theme}
+                  />
+                </Card.Content>
+
+                <Card.Content style={styles.cardContent}>
+                  <Text style={styles.label}>Confirmação de senha</Text>
+
+                  <Input
+                    value={values.confirmPassword}
+                    onChangeText={handleChange('confirmPassword')}
+                    onBlur={() => setFieldTouched('confirmPassword', true)}
+                    secureTextEntry={true}
+                    placeholder="*********"
+                    error={errors.confirmPassword && touched.confirmPassword ? true : false}
+                    theme={theme}
+                  />
+                </Card.Content>
+
+                <Card.Content style={styles.cardContent}>
+                  <View style={styles.buttonContainer}>
+                    <Button onPress={() => handleSubmit()} title={'ADICIONAR'} />
+                  </View>
+                </Card.Content>
+              </Form>
+            </React.Fragment>
+          )}
+        </Formik>
       </KeyboardAvoidingView>
 
       <Toast
@@ -117,6 +272,23 @@ const styles = StyleSheet.create({
   container: {
     backgroundColor: AppColors.BLUE,
     flex: 1,
+  },
+
+  // form leader styles
+  cardContent: {
+    marginBottom: 5,
+  },
+  label: {
+    fontFamily: 'Montserrat_medium',
+    fontSize: 12,
+  },
+  input: {
+    backgroundColor: '#ffff',
+    fontSize: 12,
+    fontFamily: 'Montserrat_medium',
+  },
+  buttonContainer: {
+    marginTop: 32,
   },
 });
